requirements.txt
# ── Core Flask ──────────────────────────────
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.7
Flask-Login==0.6.3
Flask-WTF==1.2.1
WTForms==3.1.2
Werkzeug==3.0.3

# ── Database ─────────────────────────────────
PyMySQL==1.1.1
cryptography==42.0.8

# ── Environment ──────────────────────────────
python-dotenv==1.0.1

# ── Validation ───────────────────────────────
email-validator==2.1.1

# ── Reports & Export ─────────────────────────
openpyxl==3.1.2
reportlab==4.2.2

# ── Testing ──────────────────────────────────
pytest==8.2.2
pytest-flask==1.3.0